# docs/openapi_spec.json
{
...
  "info": {
    "description": "Comprehensive API for managing forms, responses, and AI analysis.",
    "title": "Forms Backend API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/form/api/v1/workflows/{workflow_id}": {
      "delete": {
        "parameters": [
          {
            "in": "path",
            "name": "workflow_id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Soft-delete a workflow."
          }
        },
        "summary": "Soft-delete a workflow.",
        "tags": [
          "Workflow"
        ]
      },
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "workflow_id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Get detailed workflow definition."
          }
        },
        "summary": "Get detailed workflow definition.",
        "tags": [
          "Workflow"
        ]
      },
      "put": {
        "parameters": [
          {
            "in": "path",
            "name": "workflow_id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Update an existing workflow."
          }
        },
        "summary": "Update an existing workflow.",
        "tags": [
          "Workflow"
        ]
      }
    }
  },
  "securityDefinitions": {
    "Bearer": {
      "description": "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\"",
      "in": "header",
      "name": "Authorization",
      "type": "apiKey"
    }
  },
  "swagger": "2.0"
}